+++ index.ts
@@ -3,5 +3,5 @@ export { default as buildDataUri } from './src/buildDataUri/buildDataUri';
 export type { QueryObject } from './src/buildSearchParams/buildSearchParams';
 export { default as buildSearchParams } from './src/buildSearchParams/buildSearchParams';
 export { default as buildUrl } from './src/buildUrl/buildUrl';
-export type { Stringifiable } from './src/SafeUrl/SafeUrl';
-export { default as SafeUrl } from './src/SafeUrl/SafeUrl';
+export type { Stringifiable } from './src/SafeURL/SafeURL';
+export { default as SafeURL } from './src/SafeURL/SafeURL';