+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 1.1.0 - Aug 30, 2026
 
 - buildSearchParams function
+- buildDataUrl function
+- SafeURL class
 
 ## 1.0.0 - Aug 29, 2026